backend/requirements.txt
# Backend Python dependencies for AutoMate
# Pin or loosen versions as needed for your environment. These were inferred from imports
# in the backend Django project (Django 5.2 was referenced in settings.py).

Django>=5.2,<6
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2   # only if you plan to use JWT (not required by current code)
django-cors-headers>=4.0
django-filter>=24.0
drf-yasg>=1.21
google-cloud-vision>=3.0.0
google-auth>=2.20
python-dotenv>=1.0
Pillow>=10.0
openai>=2.0.0
requests>=2.28.0

# Task scheduling and background jobs
celery>=5.3.0
redis>=4.5.0
django-celery-beat>=2.5.0

# PostgreSQL adapter (required for production)
psycopg2-binary>=2.9

# Optional / common extras
ipython

# Production server and static files
gunicorn>=21.0
whitenoise>=6.0

# Pin the test runner if you use pytest
pytest>=7.0
pytest-django>=4.5

# Error tracking (optional but recommended)
sentry-sdk>=1.40.0

# AWS S3 storage (optional, for media files)
# boto3>=1.34.0
# django-storages>=1.14.0

